Move date functions into helper
